pyproject.toml
[tool.poetry]
name = "cxexam-spider"
version = "0.1.0"
description = "ChaoXing course video&exam auto flasher"
authors = ["SocialSisterYi"]
license = "GPL-3.0"
readme = "README.md"
packages = [{include = "cxexam_spider"}]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true


[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
maskpass = "^0.3.7"
qrcode = "^7.3.1"
rich = "^12.6.0"
pycryptodome = "^3.15.0"
lxml = "^4.9.1"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
